server-side calls v10, v2
