+++ packages/server/src/core/index.ts
@@ -2,5 +2,4 @@ export type { AnyRouter, ProcedureRouterRecord } from './router';
 export type { Procedure, ProcedureParams, ProcedureArgs } from './procedure';
 
 export { initTRPC } from './initTRPC';
-export { createCaller } from './caller';
 export * from './types';
